Second Sprint/requirements.txt
streamlit==1.28.1
tensorflow==2.13.0
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
matplotlib==3.7.2